Adversy/requirements.txt
requests
python-dotenv
rich
vaderSentiment